Google Maps

+++ app/src/main/java/co/edu/eam/fakemaps/modelo/Posicion.kt
@@ -5,7 +5,7 @@ class Posicion (){
     var lat:Double = 0.0
     var lng:Double = 0.0
 
-    constructor(lat: Float, lng: Float):this(){
+    constructor(lat: Double, lng: Double):this(){
         this.lat = lat.toDouble()
         this.lng = lng.toDouble()
     }
